requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
numpy>=1.21.0
pandas>=1.3.0
Pillow>=9.0.0
PyYAML>=5.4.0
matplotlib>=3.4.0
scikit-image>=0.18.0
nltk>=3.6.0
pycocoevalcap>=1.2
captum>=0.5.0
tensorboard>=2.8.0
tqdm>=4.62.0
scipy>=1.7.0
opencv-python>=4.5.0
requests>=2.26.0
